requirements.txt
# This file lists all dependencies for the project.
# Use `pip install -r requirements.txt` to install all the listed packages.
openai
wxPython
python-dotenv